+++ src/delimiters.ts
@@ -7,13 +7,10 @@ export class Delimiters {
     public containerTagOpen = "#";
     public containerTagClose = "/";
 
-    constructor(initial?: Delimiters) {
+    constructor(initial?: Partial<Delimiters>) {
         Object.assign(this, initial);
 
         this.encodeAndValidate();
-
-        if (this.tagStart === this.tagEnd)
-            throw new Error(`${nameof(this.tagStart)} can not be equal to ${nameof(this.tagEnd)}`);
 
         if (this.containerTagOpen === this.containerTagClose)
             throw new Error(`${nameof(this.containerTagOpen)} can not be equal to ${nameof(this.containerTagClose)}`);
@@ -26,9 +23,6 @@ export class Delimiters {
 
             if (!value)
                 throw new Error(`${key} must be specified.`);
-
-            if (value.length > 1)
-                throw new Error(`Only single character delimiters supported (${key}: '${value}').`);
 
             this[key] = XmlNode.encodeValue(value);
         }