+++ src/utils/array.ts
@@ -5,6 +5,12 @@ export function pushMany<T>(destArray: T[], items: T[]): void {
     Array.prototype.push.apply(destArray, items);
 }
 
+export function first<T>(array: T[]): T {
+    if (!array.length)
+        return undefined;
+    return array[0];
+}
+
 export function last<T>(array: T[]): T {
     if (!array.length)
         return undefined;