+++ src/xml/index.ts
@@ -1,2 +1,3 @@
+export * from './xmlDepthTracker';
 export * from './xmlNode';
 export * from './xmlParser';